Servlet rights check on mobile_modules by name are moved to a lower level FocObjectServlet
